+++ lib/config.dart
@@ -88,6 +88,11 @@ class Config {
   /// Intended to be used in E2E testing.
   static bool disableDragArea = false;
 
+  /// Indicator whether any activity in [AppLifecycleState.detached] is allowed.
+  ///
+  /// Intended to be used in E2E testing.
+  static bool allowDetachedActivity = false;
+
   /// Product identifier of `User-Agent` header to put in network queries.
   static String userAgentProduct = 'Gapopa';
 
@@ -151,6 +156,11 @@ class Config {
 
   /// URL address of IP address discovering API server.
   static String ipEndpoint = 'https://api.ipify.org?format=json';
+
+  /// Map of localized [Announcement]s that should be displayed within the app.
+  ///
+  /// Each key is expected to be a string of Unicode BCP47 Locale Identifier.
+  static Map<String, Announcement> announcements = {};
 
   /// Initializes this [Config] by applying values from the following sources
   /// (in the following order):
@@ -277,6 +287,39 @@ class Config {
         ? const String.fromEnvironment('SOCAPP_IP_ENDPOINT')
         : (document['ip']?['endpoint'] ?? ipEndpoint);
 
+    try {
+      final dynamic announcementsOrNull = document['announcement'];
+      if (announcementsOrNull is Map<String, dynamic>) {
+        announcements.clear();
+        for (var entry in announcementsOrNull.entries) {
+          announcements[entry.key] = Announcement.parse(entry.value);
+        }
+      }
+
+      // `bool.hasEnvironment` can only be used as a `const` constructor.
+      if (const bool.hasEnvironment('SOCAPP_ANNOUNCEMENT_EN_US')) {
+        announcements['en-US'] = Announcement.parse(
+          const String.fromEnvironment('SOCAPP_ANNOUNCEMENT_EN_US'),
+        );
+      }
+
+      // `bool.hasEnvironment` can only be used as a `const` constructor.
+      if (const bool.hasEnvironment('SOCAPP_ANNOUNCEMENT_ES_ES')) {
+        announcements['es-ES'] = Announcement.parse(
+          const String.fromEnvironment('SOCAPP_ANNOUNCEMENT_ES_ES'),
+        );
+      }
+
+      // `bool.hasEnvironment` can only be used as a `const` constructor.
+      if (const bool.hasEnvironment('SOCAPP_ANNOUNCEMENT_RU_RU')) {
+        announcements['en-US'] = Announcement.parse(
+          const String.fromEnvironment('SOCAPP_ANNOUNCEMENT_RU_RU'),
+        );
+      }
+    } catch (e) {
+      Log.warning('init() -> announcements failed: $e', 'Config');
+    }
+
     // Change default values to browser's location on web platform.
     if (PlatformUtils.isWeb) {
       if (document['server']?['http']?['url'] == null &&
@@ -315,7 +358,7 @@ class Config {
     if (confRemote) {
       try {
         final response = await (await PlatformUtils.dio).fetch(
-          RequestOptions(path: '$url:$port/conf'),
+          RequestOptions(path: '$url:$port/conf?${Pubspec.ref}'),
         );
         if (response.statusCode == 200) {
           dynamic remote;
@@ -359,6 +402,21 @@ class Config {
             }
             logAmount = _asInt(remote['log']?['amount']) ?? logAmount;
             logObfuscated = remote['log']?['obfuscated'] ?? logObfuscated;
+
+            try {
+              final dynamic announcementsOrNull = remote['announcement'];
+              if (announcementsOrNull is Map) {
+                announcements.clear();
+                for (var entry in announcementsOrNull.entries) {
+                  if (entry.key is String && entry.value is String) {
+                    announcements[entry.key] = Announcement.parse(entry.value);
+                  }
+                }
+              }
+            } catch (e) {
+              Log.warning('init() -> announcements failed: $e', 'Config');
+            }
+
             origin = url;
           }
         }
@@ -395,6 +453,38 @@ class Config {
       });
     }
   }
+}
+
+/// [title] with a [body] intended to be used as an announcement to make to the
+/// client in a non-dismissible way.
+class Announcement {
+  const Announcement({required this.title, this.body});
+
+  /// Constructs an [Announcement] from the provided [text].
+  ///
+  /// First line of the [text], if there's multiple present, is considered as a
+  /// [title].
+  factory Announcement.parse(String text) {
+    final split = text.split('\n');
+    if (split.length >= 2) {
+      final String title = split.first;
+      return Announcement(
+        title: title.trim(),
+        body: text.substring(title.length + 1).trim(),
+      );
+    }
+
+    return Announcement(title: text);
+  }
+
+  /// Indicates whether this [Announcement] is empty.
+  bool get isEmpty => title.isEmpty && body?.isNotEmpty != true;
+
+  /// Title of this [Announcement].
+  final String title;
+
+  /// Optional body of this [Announcement].
+  final String? body;
 }
 
 /// Parses the provided [val] as [int].