+++ lib/routes.dart
@@ -655,6 +655,19 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
                     ),
                   );
 
+              final AbstractSessionRepository sessionRepository = deps
+                  .put<AbstractSessionRepository>(
+                    SessionRepository(
+                      graphQlProvider,
+                      Get.find(),
+                      Get.find(),
+                      Get.find(),
+                      Get.find(),
+                      Get.find(),
+                    ),
+                  );
+              deps.put<SessionService>(SessionService(sessionRepository));
+
               deps.put(MyUserService(Get.find(), myUserRepository));
               deps.put(UserService(userRepository));
               deps.put(ContactService(contactRepository));
@@ -1144,7 +1157,10 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
   Widget build(BuildContext context) {
     return SentryDisplayWidget(
       child: LifecycleObserver(
-        onStateChange: (v) => _state.lifecycle.value = v,
+        onStateChange: (v) {
+          Log.debug('onStateChange() -> $v', '$runtimeType');
+          _state.lifecycle.value = v;
+        },
         child: Listener(
           onPointerDown: (_) => PlatformUtils.keepActive(),
           onPointerHover: (_) => PlatformUtils.keepActive(),