+++ lib/ui/page/home/page/chat/message_field/controller.dart
@@ -71,9 +71,7 @@ class MessageFieldController extends GetxController {
        ) {
     field = TextFieldState(
       text: text,
-      onFocus: (s) {
-        onChanged?.call();
-      },
+      onFocus: (s) => onChanged?.call(),
       submitted: false,
       onSubmitted: (s) {
         field.unsubmit();
@@ -569,6 +567,11 @@ class MessageFieldController extends GetxController {
   ///
   /// May be used to test a [file] upload since [FilePicker] can't be mocked.
   Future<void> _addAttachment(NativeFile file) async {
+    Log.debug(
+      '_addAttachment($file) -> edited(${edited.value})',
+      '$runtimeType',
+    );
+
     if (file.size < maxAttachmentSize && _chatService != null) {
       try {
         var attachment = LocalAttachment(file, status: SendingStatus.sending);