+++ lib/ui/page/home/page/chat/view.dart
@@ -1340,7 +1340,6 @@ class ChatView extends StatelessWidget {
                   sizeFactor: animation,
                   axis: Axis.horizontal,
                   axisAlignment: 0,
-
                   child: ScaleTransition(
                     scale: animation,
                     alignment: Alignment.centerLeft,