+++ lib/ui/page/login/view.dart
@@ -100,7 +100,9 @@ class LoginView extends StatelessWidget {
                 ReactiveTextField(
                   key: const Key('RecoveryField'),
                   state: c.recovery,
-                  label: 'label_sign_in_input'.l10n,
+                  label: 'label_identifier'.l10n,
+                  hint: 'label_sign_in_input'.l10n,
+                  floatingLabelBehavior: FloatingLabelBehavior.always,
                 ),
                 const SizedBox(height: 25),
                 PrimaryButton(
@@ -154,32 +156,22 @@ class LoginView extends StatelessWidget {
                   style: style.fonts.normal.regular.secondary,
                 ),
                 const SizedBox(height: 25),
-                ReactiveTextField(
+                ReactiveTextField.password(
                   key: const Key('PasswordField'),
                   state: c.newPassword,
                   label: 'label_new_password'.l10n,
-                  obscure: c.obscureNewPassword.value,
-                  onSuffixPressed: c.obscureNewPassword.toggle,
+                  hint: 'label_enter_password'.l10n,
+                  obscured: c.obscureNewPassword,
                   treatErrorAsStatus: false,
-                  trailing: SvgIcon(
-                    c.obscureNewPassword.value
-                        ? SvgIcons.visibleOff
-                        : SvgIcons.visibleOn,
-                  ),
                 ),
                 const SizedBox(height: 16),
-                ReactiveTextField(
+                ReactiveTextField.password(
                   key: const Key('RepeatPasswordField'),
                   state: c.repeatPassword,
-                  label: 'label_repeat_password'.l10n,
-                  obscure: c.obscureRepeatPassword.value,
-                  onSuffixPressed: c.obscureRepeatPassword.toggle,
+                  label: 'label_confirm_password'.l10n,
+                  hint: 'label_repeat_password'.l10n,
+                  obscured: c.obscureRepeatPassword,
                   treatErrorAsStatus: false,
-                  trailing: SvgIcon(
-                    c.obscureRepeatPassword.value
-                        ? SvgIcons.visibleOff
-                        : SvgIcons.visibleOn,
-                  ),
                 ),
                 const SizedBox(height: 25),
                 Obx(() {
@@ -212,36 +204,22 @@ class LoginView extends StatelessWidget {
                   floatingLabelBehavior: FloatingLabelBehavior.always,
                 ),
                 const SizedBox(height: 16),
-                ReactiveTextField(
+                ReactiveTextField.password(
                   key: const ValueKey('PasswordField'),
                   state: c.password,
                   label: 'label_password'.l10n,
-                  obscure: c.obscurePassword.value,
-                  onSuffixPressed: c.obscurePassword.toggle,
+                  hint: 'label_enter_password'.l10n,
+                  obscured: c.obscurePassword,
                   treatErrorAsStatus: false,
-                  trailing: SvgIcon(
-                    c.obscurePassword.value
-                        ? SvgIcons.visibleOff
-                        : SvgIcons.visibleOn,
-                  ),
-                  hint: '***',
-                  floatingLabelBehavior: FloatingLabelBehavior.always,
                 ),
                 const SizedBox(height: 16),
-                ReactiveTextField(
+                ReactiveTextField.password(
                   key: const ValueKey('RepeatPasswordField'),
                   state: c.repeatPassword,
-                  label: 'label_password'.l10n,
-                  obscure: c.obscureRepeatPassword.value,
-                  onSuffixPressed: c.obscureRepeatPassword.toggle,
+                  label: 'label_confirm_password'.l10n,
+                  hint: 'label_repeat_password'.l10n,
+                  obscured: c.obscureRepeatPassword,
                   treatErrorAsStatus: false,
-                  trailing: SvgIcon(
-                    c.obscureRepeatPassword.value
-                        ? SvgIcons.visibleOff
-                        : SvgIcons.visibleOn,
-                  ),
-                  hint: '***',
-                  floatingLabelBehavior: FloatingLabelBehavior.always,
                 ),
                 const SizedBox(height: 25),
                 Obx(() {
@@ -400,21 +378,18 @@ class LoginView extends StatelessWidget {
                 ReactiveTextField(
                   key: const Key('UsernameField'),
                   state: c.login,
-                  label: 'label_sign_in_input'.l10n,
+                  label: 'label_identifier'.l10n,
+                  hint: 'label_sign_in_input'.l10n,
+                  floatingLabelBehavior: FloatingLabelBehavior.always,
                 ),
                 const SizedBox(height: 16),
-                ReactiveTextField(
+                ReactiveTextField.password(
                   key: const ValueKey('PasswordField'),
                   state: c.password,
                   label: 'label_password'.l10n,
-                  obscure: c.obscurePassword.value,
-                  onSuffixPressed: c.obscurePassword.toggle,
+                  hint: 'label_your_password'.l10n,
+                  obscured: c.obscurePassword,
                   treatErrorAsStatus: false,
-                  trailing: SvgIcon(
-                    c.obscurePassword.value
-                        ? SvgIcons.visibleOff
-                        : SvgIcons.visibleOn,
-                  ),
                 ),
                 Padding(
                   padding: const EdgeInsets.fromLTRB(21, 8, 8, 8),